+++ shotmanager/handlers/sm_handlers.py
@@ -19,6 +19,9 @@
 Handler functions
 """
 
+import bpy
+from bpy.app.handlers import persistent
+
 import logging
 
 _logger = logging.getLogger(__name__)
@@ -27,6 +30,9 @@ _logger = logging.getLogger(__name__)
 # @persistent
 def shotMngHandler_undo_pre(self, context):
     print("\nSM Handler: Undo Pre")
+
+
+#   bpy.context.window_manager.UAS_shot_manager_display_overlay_tools = False
 
 
 # @persistent
@@ -38,5 +44,11 @@ def shotMngHandler_undo_post(self, context):
 def shotMngHandler_load_pre(self, context):
     print("\nSM Handler: Load Pre")
 
-# context.window_manager.UAS_shot_manager_display_overlay_tools = False
+    # bpy.ops.uas_shot_manager.sequence_timeline.cancel(bpy.context)
+
+    # bpy.context.window_manager.event_timer_remove(bpy.ops.uas_shot_manager.sequence_timeline.draw_event)
+    # bpy.types.SpaceView3D.draw_handler_remove(bpy.ops.uas_shot_manager.sequence_timeline.draw_handle, "WINDOW")
+
+    # bpy.ops.uas_shot_manager.sequence_timeline.unregister_handlers(context)
+    # bpy.context.window_manager.UAS_shot_manager_display_overlay_tools = False